+++ types/index.test-d.ts
@@ -41,7 +41,7 @@ import remove, {
   removeOrThrowIfNotExists,
 } from './remove';
 import reverseIterate from './reverseIterate';
-import split from './split';
+import split, {type SplitResult} from './split';
 import toArray from './toArray';
 import union from './union';
 import update, {
@@ -109,7 +109,7 @@ expectType<types.ImmutableTree<string>>(map<number, string>(numberTree, toString
 expectType<types.ImmutableTree<string>>(remove<string, string>(stringTree, '', cmpStrings));
 expectType<types.ImmutableTree<string>>(removeIfExists<string, string>(stringTree, '', cmpStrings));
 expectType<types.ImmutableTree<string>>(removeOrThrowIfNotExists<string, string>(stringTree, '', cmpStrings));
-expectType<[types.ImmutableTree<number>, types.ImmutableTree<number>, types.ImmutableTree<number>]>(split(numberTree, 0, cmpNumbers));
+expectType<SplitResult<number>>(split(numberTree, 0, cmpNumbers));
 expectType<Array<string>>(toArray(stringTree));
 expectType<types.ImmutableTree<string>>(union(stringTree, stringTree, cmpStrings));
 expectType<types.ImmutableTree<string>>(union(stringTree, stringTree, cmpStrings, (v1, v2) => v2));
@@ -127,7 +127,7 @@ expectType<string | null>(findNext<string, number, null>(stringTree, 0, cmpNumbe
 expectType<string | null>(findPrev<string, number, null>(stringTree, 0, cmpNumberAndString, null));
 expectType<[string | null, number]>(findWithIndex<string, number, null>(stringTree, 0, cmpNumberAndString, null));
 expectType<number>(indexOf<string, number>(stringTree, 1, cmpNumberAndString));
-expectType<[types.ImmutableTree<string>, types.ImmutableTree<string>, types.ImmutableTree<string>]>(split<string, number>(stringTree, 1, cmpNumberAndString));
+expectType<SplitResult<string>>(split<string, number>(stringTree, 1, cmpNumberAndString));
 
 // Wrong tree type
 expectError<types.ImmutableTree<number>>(create<number>(''));
@@ -149,7 +149,7 @@ expectError<number>(minValue<number>(stringTree));
 expectError<types.ImmutableTree<number>>(remove<number, number>(stringTree, 0, cmpNumbers));
 expectError<types.ImmutableTree<number>>(removeIfExists<number, number>(stringTree, 0, cmpNumbers));
 expectError<types.ImmutableTree<number>>(removeOrThrowIfNotExists<number, number>(stringTree, 0, cmpNumbers));
-expectError<[types.ImmutableTree<number>, types.ImmutableTree<number>, types.ImmutableTree<number>]>(split<number>(stringTree, 1, cmpNumbers));
+expectError<SplitResult<number>>(split<number>(stringTree, 1, cmpNumbers));
 expectError<ReadonlyArray<string>>(toArray<string>(numberTree));
 expectError<types.ImmutableTree<number>>(union<number>(stringTree, stringTree, cmpNumbers));
 expectError<types.ImmutableTree<number>>(difference<number>(stringTree, stringTree, cmpNumbers));
@@ -167,7 +167,7 @@ expectError<types.ImmutableTree<string>>(insert<string>(stringTree, '', cmpNumbe
 expectError<types.ImmutableTree<string>>(remove<string, string>(stringTree, '', cmpNumbers));
 expectError<types.ImmutableTree<string>>(removeIfExists<string, string>(stringTree, '', cmpNumbers));
 expectError<types.ImmutableTree<string>>(removeOrThrowIfNotExists<string, string>(stringTree, '', cmpNumbers));
-expectError<[types.ImmutableTree<string>, types.ImmutableTree<string>, types.ImmutableTree<string>]>(split<string>(stringTree, '', cmpNumbers));
+expectError<SplitResult<string>>(split<string>(stringTree, '', cmpNumbers));
 expectError<types.ImmutableTree<string>>(union<string>(stringTree, stringTree, cmpNumbers));
 expectError<types.ImmutableTree<string>>(difference<string>(stringTree, stringTree, cmpNumbers));
 expectError<types.ImmutableTree<string>>(intersection<number>(stringTree, stringTree, cmpNumbers));
@@ -223,7 +223,7 @@ expectType<types.ImmutableTree<NSTuple>>(mapTreeWrapper.intersection(numberStrin
 expectType<types.ImmutableTree<NSTuple>>(mapTreeWrapper.remove(numberStringMapTree, [0, ''] as NSTuple));
 expectType<types.ImmutableTree<NSTuple>>(mapTreeWrapper.removeIfExists(numberStringMapTree, [0, ''] as NSTuple));
 expectType<types.ImmutableTree<NSTuple>>(mapTreeWrapper.removeOrThrowIfNotExists(numberStringMapTree, [0, ''] as NSTuple));
-expectType<[types.ImmutableTree<NSTuple>, types.ImmutableTree<NSTuple>, types.ImmutableTree<NSTuple>]>(mapTreeWrapper.split(numberStringMapTree, 0));
+expectType<SplitResult<NSTuple>>(mapTreeWrapper.split(numberStringMapTree, 0));
 expectType<types.ImmutableTree<NSTuple>>(mapTreeWrapper.union(numberStringMapTree, numberStringMapTree));
 expectType<types.ImmutableTree<NSTuple>>(mapTreeWrapper.union(numberStringMapTree, numberStringMapTree, onConflictThrowError));
 expectType<types.ImmutableTree<NSTuple>>(mapTreeWrapper.update(numberStringMapTree, 1, onConflictThrowError, onNotFoundUseGivenValue));