+++ types/split.d.ts
@@ -1,11 +1,13 @@
 import type {ImmutableTree} from './types';
 
-export default function split<T, K = T>(
-  tree: ImmutableTree<T>,
-  key: K,
-  cmp: (a: K, b: T) => number,
-): [
+export type SplitResult<T> = readonly [
   small: ImmutableTree<T>,
   equal: ImmutableTree<T>,
   large: ImmutableTree<T>,
 ];
+
+export default function split<T, K = T>(
+  tree: ImmutableTree<T>,
+  key: K,
+  cmp: (a: K, b: T) => number,
+): SplitResult<T>;